Refactor role -> variant
In the beginning I thought there would be many more kinds of images than
server and workstation, but over time everything I need sort of fits
into one of the two categories, making it less descriptive as a host's
function. Rename to avoid confusion with other uses of the word, such as
in the Puppet roles and profiles pattern, and Kubernetes node roles.