bootloader: Fix sysctl.vm.swappiness parameter
The vm.swappiness parameter was never honored by the kernel, but might
have been read by systemd or older inits. It's not working anymore, but
I want to keep this setting here next to the zswap settings its
associated with. Linux 6 allows setting sysctl parameters on the kernel
command line.