vscode: Force window reuse
Normally, VS Code decides whether to open or reuse a window based on a
setting (opening file vs. folder), but the args added by this wrapper
make it think it should always open a new window. That's almost never
what I want, so force the expected behavior with an additional arg.