systemd: Work around race condition between resolved and networkd
On Milk-V Pioneer, the network interface comes up but not online during
which time systemd-resolved decides to downgrade the nameserver features
such as DNSSEC, blocking further requests until it's restarted or reset.
Modify the wait-online service to only wait for 'any' interface on
systems with wifi (the original intent) and add a service to reset
resolver features when all configured interfaces are up and online.