systemd: Add sleep to resolved reset
Restore the previous networkd-wait-online behavior and just add a delay
for resetting nameserver features. This race condition is proving to be
tricky, because systemd-resolved initiates nameserver feature checks
before the interface is ready, and simply waiting for the interface to
come online may not catch the fetaure check at the right time.