fstab: Make relationship between NFS mounts and VPN stronger
The openvpn-client@.service seems to be marked running before it's fully
up, so services that attempt to access these mountpoints like /nest
ultimately fail at boot. Introducing an additional dependency on the VPN
interface seems to take care of it, though a proper solution would be in
the openvpn service itself.