k8s: Restore LoadBalancer externalTrafficPolicy
After upgrading to kubernetes 1.34.1, the bug blocking traffic to
services with `externalTrafficPolicy: Local` is fixed. kube-proxy's
nftables backend is still owning the routes, so external traffic is
still blocked to these services, but we will solve that by moving
external routing to a host that is not also a k8s node and thus does not
have kube-proxy. BGP will facilitate the routing as it always has.

Enable the 'Local' policy for jump services, too, such that SSH now
records the last login IP correctly and traffic isn't NATed.

Closes #72