gitlab: Probably working incoming email config
This feature is really hard to test, because it really demands multiple
users and email accounts, which I don't really have. This is copied from
a once working omnibus config. Gmail shows an active IMAP connection, so
it might be working. Again, hard to test, and not something I *ever*
used, so I'm dropping it.