raspberrypi: Try upstream kernel
Having all sorts of trouble getting the mmc driver to work, which this
issue, https://github.com/raspberrypi/linux/issues/7025, suggest may be
related to the u-boot handoff to Raspberry Pi Linux, or in my case, the
u-boot to systemd-boot to Linux handoff. It suggests upstream linux
works better with u-boot.